pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "symphony"
version = "0.1.0"
description = "Symphony orchestrates coding agents against issue tracker work."
readme = "README.md"
requires-python = ">=3.9"
license = { text = "Apache-2.0" }
dependencies = [
    "PyYAML>=6.0",
    "httpx>=0.25",
    "Jinja2>=3.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7",
    "pytest-asyncio>=0.21",
]

[project.scripts]
symphony = "symphony.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["symphony*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
